+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ## [Unreleased]
 ### Changed
 - `JSON`: modified error messages (removed "at root" when pointer is at root of JSON)
+- `JSON`: added `escape` and `unescape`
 
 ## [4.2] - 2020-12-26
 ### Changed